Rename create-ref to input-ref
